requirements/base.txt
# Django
Django==1.8.9

# Environmental var configuration
django-configurations==1.0
envdir==0.7

# For the persistence stores
psycopg2==2.6.1
dj-database-url==0.3.0

# Models
django-model-utils==2.4

# Rest apis
djangorestframework==3.3.2
Markdown==2.6.5
django-filter==0.11.0
requests==2.9.1

# Time zones support
pytz==2015.7
